+++ common/src/main/java/net/satisfy/bakery/core/network/PacketHandler.java
@@ -4,6 +4,7 @@ import dev.architectury.networking.NetworkManager;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
 import net.satisfy.bakery.Bakery;
+import net.satisfy.bakery.core.network.packet.SetStreetSignTextPacket;
 
 public class PacketHandler {
     public static final ResourceLocation SET_SIGN_TEXT = Bakery.identifier( "set_sign_text");