+++ common/src/main/java/net/satisfy/bakery/core/item/SugarRushEffectItem.java
@@ -1,5 +1,6 @@
 package net.satisfy.bakery.core.item;
 
+import dev.architectury.registry.registries.RegistrySupplier;
 import net.minecraft.core.Holder;
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectInstance;
@@ -7,16 +8,17 @@ import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
+import net.satisfy.bakery.core.registry.MobEffectRegistry;
 import net.satisfy.farm_and_charm.core.item.food.EffectItem;
 import org.jetbrains.annotations.NotNull;
 
 public class SugarRushEffectItem extends EffectItem {
 
-    private final Holder<MobEffect> effect;
+    private final RegistrySupplier<MobEffect> effect;
     private final int duration;
     private final int maxStacks;
 
-    public SugarRushEffectItem(Properties properties, Holder<MobEffect> effect, int duration, int maxStacks, boolean returnBowl) {
+    public SugarRushEffectItem(Properties properties, RegistrySupplier<MobEffect> effect, int duration, int maxStacks, boolean returnBowl) {
         super(properties, duration, returnBowl);
         this.effect = effect;
         this.duration = duration;
@@ -35,16 +37,12 @@ public class SugarRushEffectItem extends EffectItem {
             return result;
         }
 
-        MobEffectInstance currentEffect = player.getEffect(effect);
+        Holder<MobEffect> effectHolder = MobEffectRegistry.holder(effect);
+        MobEffectInstance currentEffect = player.getEffect(effectHolder);
 
-        int newAmplifier;
-        if (currentEffect == null) {
-            newAmplifier = 0;
-        } else {
-            newAmplifier = Math.min(maxStacks - 1, currentEffect.getAmplifier() + 1);
-        }
+        int newAmplifier = currentEffect == null ? 0 : Math.min(maxStacks - 1, currentEffect.getAmplifier() + 1);
 
-        player.addEffect(new MobEffectInstance(effect, duration, newAmplifier, false, true, true));
+        player.addEffect(new MobEffectInstance(effectHolder, duration, newAmplifier, false, true, true));
 
         return result;
     }